Update on review and add more tests
